chore: Rename contentful services file and change to typescript
